Add rubocop
