+++ .github/workflows/ci.yaml
@@ -28,3 +28,5 @@ jobs:
         bundler-cache: true
     - name: Run tests
       run: bundle exec rake
+    - name: Rubocop
+      run: bundle exec rubocop
