+++ .rubocop.yml
@@ -1,0 +1,65 @@
+AllCops:
+  NewCops: enable
+
+Layout/SpaceInsideParens:
+  Enabled: false
+
+Layout/MultilineArrayBraceLayout:
+  Enabled: false
+
+Metrics/ClassLength:
+  Enabled: false
+
+Metrics/AbcSize:
+  Enabled: false
+
+Metrics/CyclomaticComplexity:
+  Enabled: false
+
+Metrics/MethodLength:
+  Enabled: false
+
+Metrics/ParameterLists:
+  Enabled: false
+
+Metrics/PerceivedComplexity:
+  Enabled: false
+
+Naming/AccessorMethodName:
+  Exclude:
+    - 'lib/minitest/proptest.rb'
+
+Naming/MethodParameterName:
+  Enabled: false
+
+Style/Documentation:
+  Exclude:
+    - 'lib/minitest/proptest.rb'
+    - 'lib/minitest/proptest_plugin.rb'
+    - 'test/**/*'
+
+Style/EachWithObject:
+  Enabled: false
+
+Style/EmptyLambdaParameter:
+  Enabled: false
+
+Style/Lambda:
+  Enabled: false
+
+Style/MultilineBlockChain:
+  Enabled: false
+
+Style/NumericPredicate:
+  Enabled: false
+
+Style/RedundantSelf:
+  Enabled: false
+
+Style/RescueStandardError:
+  Exclude:
+    - 'lib/minitest/proptest/property.rb'
+
+Style/SoleNestedConditional:
+  Exclude:
+    - 'lib/minitest/proptest/property.rb'
